# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Organisation Registry Backend API",
    "description": "Organisation Registry Backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/node/{id}/attributes/names": {
      "put": {
        "summary": "PUT api/node/{id}/attributes/names",
        "operationId": "updateNameAttributes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "organisation unit unique id",
            "example": "59457884"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NameAttribute"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEntity"
                }
              }
            }
          },
          "422": {
            "description": "Bad request. Validation error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseEntity"
                },
                "examples": {
                  "validation error no value" : {
                    "$ref": "#/components/examples/ex1"
                  },
                  "validation error in dates" : {
                    "$ref":"#/components/examples/ex2"
                  },
                  "validation error value too long" : {
                    "$ref":"#/components/examples/ex3"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "GET api/node/{id}/attributes/names",
        "operationId": "getNodeNameAttributes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "organisation unit unique id",
            "example": "59457884"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/node/{id}/attributes/types": {
      "put": {
        "summary": "PUT api/node/{id}/attributes/types",
        "operationId": "updateTypeAttributes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "organisation unit unique id",
            "example": "59457884"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TypeAttribute"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEntity"
                }
              }
            }
          },
          "422": {
            "description": "Bad request. Validation error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseEntity"
                },
                "examples": {
                  "validation error no value" : {
                    "$ref": "#/components/examples/ex1"
                  },
                  "validation error in dates" : {
                    "$ref":"#/components/examples/ex2"
                  },
                  "validation error value too long" : {
                    "$ref":"#/components/examples/ex3"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "GET api/node/{id}/attributes/types",
        "operationId": "getNodeTypeAttributes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "organisation unit unique id",
            "example": "59457884"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypeAttribute"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/node/{id}/attributes/codes": {
      "put": {
        "summary": "PUT api/node/{id}/attributes/codes",
        "operationId": "updateCodeAttributes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "organisation unit unique id",
            "example": "59457884"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CodeAttribute"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEntity"
                }
              }
            }
          },
          "422": {
            "description": "Bad request. Validation error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseEntity"
                },
                "examples": {
                  "validation error no value" : {
                    "$ref": "#/components/examples/ex1"
                  },
                  "validation error in dates" : {
                    "$ref":"#/components/examples/ex2"
                  },
                  "validation error value too long" : {
                    "$ref":"#/components/examples/ex3"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "GET api/node/{id}/attributes/codes",
        "operationId": "getNodeCodeAttributes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "organisation unit unique id",
            "example": "59457884"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CodeAttribute"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "examples": {
      "ex1": {
        "value": {
          "id": 32032,
          "nodeId": "5283",
          "errorMessage": "error_value_is_mandatory"
        }
      },
      "ex2" : {
        "value": {
          "id": 32032,
          "nodeId": "5283",
          "errorMessage": "attribute_start_date_after_end_date"
        }
      },
      "ex3" : {
        "value": {
          "id": 32032,
          "nodeId": "5283",
          "errorMessage": "attribute_value_is_too_long"
        }
      }
    },
    "schemas": {
      "NameAttribute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 32032,
            "description": "attribute unique id",
            "required": true
          },
          "key": {
            "type": "string",
            "example": "name_fi",
            "description": "attribute language key",
            "required": true
          },
          "value": {
            "type": "string",
            "example": "Forest Economics and  Management",
            "description": "value of the attribute",
            "required": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "1800-12-02T00:00:00.000+00:00",
            "description": "start date of the attribute"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "null",
            "description": "end date of the attribute"
          },
          "isNew" : {
            "type" : "boolean",
            "example": false,
            "description": "is new attribute"
          },
          "deleted" : {
            "type" : "boolean",
            "example": false,
            "description": "is attribute marked for deletion"
          }
        }
      },
      "TypeAttribute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 5289,
            "description": "attribute unique id",
            "required": true
          },
          "key": {
            "type": "string",
            "example": "type",
            "description": "attribute type",
            "required": true
          },
          "value": {
            "type": "string",
            "example": "tulosyksikko",
            "description": "value of the attribute",
            "required": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "1800-12-02T00:00:00.000+00:00",
            "description": "start date of the attribute"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "null",
            "description": "end date of the attribute"
          },
          "isNew" : {
            "type" : "boolean",
            "example": false,
            "description": "is new attribute"
          },
          "deleted" : {
            "type" : "boolean",
            "example": false,
            "description": "is attribute marked for deletion"
          }
        }
      },
      "CodeAttribute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 5288,
            "description": "attribute unique id",
            "required": true
          },
          "key": {
            "type": "string",
            "example": "talous_tunnus",
            "description": "attribute type",
            "required": true
          },
          "value": {
            "type": "string",
            "example": "H9070",
            "description": "value of the attribute",
            "required": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "1800-12-02T00:00:00.000+00:00",
            "description": "start date of the attribute"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "null",
            "description": "end date of the attribute"
          },
          "isNew" : {
            "type" : "boolean",
            "example": false,
            "description": "is new attribute"
          },
          "deleted" : {
            "type" : "boolean",
            "example": false,
            "description": "is attribute marked for deletion"
          }
        }
      },
      "ErrorResponseNameAttribute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 32032,
            "description": "attribute unique id",
            "required": true
          },
          "nodeId": {
            "type": "string",
            "example": "5283",
            "description": "organisation unit id",
            "required": true
          },
          "errorMessage" : {
            "type": "string",
            "example": "error_name_is_mandatory",
            "description": "error text key ready to be translated"
          }
        }
      },
      "SuccessResponseEntity" : {
        "type": "array",
        "items": {
          "type": "string"
        },
        "example": [],
        "maxItems": 0
      },
      "ErrorResponseEntity": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ErrorResponseNameAttribute"
        }
      },
      "ResponseEntity" : {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/NameAttribute"
        }
      },
...
    }
  }
}